backend/requirements.txt
fastapi==0.121.0
uvicorn[standard]==0.38.0
websockets>=12.0
python-multipart>=0.0.9
httpx==0.28.1
pydantic==2.12.3
pydantic-settings>=2.5.2
python-dotenv==1.2.1
aiofiles==25.1.0
pillow==12.0.0
qrcode[pil]>=7.4
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
openai-agents[viz]==0.6.0
mcp>=1.20.0
sse-starlette==3.0.3
openai>=1.40.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-toon>=0.1.0
langsmith>=0.2.0
google-generativeai>=0.8.0

youtube-transcript-api>=1.0.0
playwright>=1.50.0
